+++ tests/test_binary_sensor.py
@@ -43,6 +43,8 @@ if TYPE_CHECKING:
         "binary_sensor.keba_keenergy_12345678_heat_pump_has_three_phase_failure",
         "binary_sensor.keba_keenergy_12345678_heat_pump_has_vfd_failure",
         "binary_sensor.keba_keenergy_12345678_heat_pump_heat_request",
+        # heating circuit
+        "binary_sensor.keba_keenergy_12345678_heat_circuit_pump_state_1",
         # solar circuit
         "binary_sensor.keba_keenergy_12345678_solar_circuit_heat_request_1_1",
         "binary_sensor.keba_keenergy_12345678_solar_circuit_heat_request_1_2",