+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add sensor for "Return flow temperature (mixer)" to the heating circuit
 - Add device for photovoltaics [(Issue #55)][issue-55]
 - Add entities to control or monitor excess energy from photovoltaics [(Issue #55)][issue-55]
+- Add "Mixer" and "Pump status" to the heating circuit [(Issue #57)][issue-57]
 
 ### Changed
 
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [issue-46]: https://github.com/superbox-dev/keba_keenergy/issues/46
 [issue-47]: https://github.com/superbox-dev/keba_keenergy/issues/47
 [issue-55]: https://github.com/superbox-dev/keba_keenergy/issues/55
+[issue-57]: https://github.com/superbox-dev/keba_keenergy/issues/57
 
 ## [1.9.1] - 2025-04-03